pytket-mbqc-py/pyproject.toml
[tool.poetry]
name = "pytket-mbqc-py"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.12,>=3.10"
notebook = "^7.1.0"
pytket-quantinuum = "^0.29.0"
pytket_pecos = "^0.1.15"
pytest = "^8.0.1"
pytket-benchmarking = {path = "../../pytket-benchmarking"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
